# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  main.CourseInfo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      language:
        type: string
      title:
        type: string
      unitCount:
        type: integer
      updatedAt:
        type: string
    type: object
  main.CourseUnitInfo:
    properties:
      courseId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      order:
        type: integer
      title:
        type: string
      trackCount:
        type: integer
      updatedAt:
        type: string
    type: object
...
  main.ErrorResponse:
    description: API错误响应
    properties:
      error:
        description: 错误信息
        example: 未授权访问
        type: string
    type: object
...
  main.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.LoginResponse:
    description: 登录成功响应
    properties:
      message:
        description: 成功消息
        example: 登录成功
        type: string
      token:
        description: JWT令牌
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        description: 用户信息
        properties:
          email:
            description: 用户邮箱
            example: user@example.com
            type: string
          id:
            description: 用户ID
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
          name:
            description: 用户名称
            example: 测试用户
            type: string
          role:
            description: 用户角色
            example: user
            type: string
        type: object
    type: object
...
  main.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 提供语言学习音频播放器的后端API服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 语言学习音频播放器API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户凭据并返回JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户并返回JWT令牌
      parameters:
      - description: 注册信息
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/main.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: 请求无效或邮箱已被注册
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 用户注册
      tags:
      - auth
  /api/auth/validate:
    get:
      consumes:
      - application/json
      description: 验证当前用户会话状态并返回用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 无效会话
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 验证用户会话
      tags:
      - auth
  /api/courses:
    get:
      consumes:
      - application/json
      description: 获取当前用户可访问的所有课程
      produces:
      - application/json
      responses:
        "200":
          description: 课程列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户课程列表
      tags:
      - courses
  /api/courses/{courseId}:
    get:
      consumes:
      - application/json
      description: 获取指定课程的详细信息
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 课程详情
          schema:
            $ref: '#/definitions/main.CourseInfo'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取课程详情
      tags:
      - courses
  /api/courses/{courseId}/units:
    get:
      consumes:
      - application/json
      description: 获取指定课程的所有单元
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 课程单元列表
          schema:
            items:
              $ref: '#/definitions/main.CourseUnitInfo'
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取课程单元列表
      tags:
      - courses
...
  /api/courses/{courseId}/units/{unitId}/tracks:
    get:
      consumes:
      - application/json
      description: 获取指定课程单元的所有音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 音轨列表
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程或单元不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单元音轨列表
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"